hamt: preserve traversal_stack uses rust stack
this was using the lightweight stack, but this presented the following
issue: NockStack::copy() can be called during the Hamt::preserve() call,
and the lightweight stack may not be empty. NockStack::copy() presumes
that the lightweight stack is empty when it is called, and thus tried to
pop elements that preserve() had pushed onto it.

NockStack::copy() should probably assert that the lightweight stack is
empty when its called